% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CST_Start.R
\name{CST_Start}
\alias{CST_Start}
\title{CSTools data retrieval function using Start}
\usage{
CST_Start(...)
}
\arguments{
\item{...}{Parameters that are automatically forwarded to the `startR::Start` 
function. See details in `?startR::Start`.}
}
\description{
This function aggregates, subsets and retrieves sub-seasonal, seasonal, 
decadal or climate projection data from NetCDF files in a local file system 
and arranges it for easy application of the CSTools functions. It calls the 
function \code{Start} from startR, which is an R package started at BSC with
the aim to develop a tool that allows the user to automatically process large 
multidimensional distributed data sets. Then, the output is transformed into 
`s2dv_cube` object.
}
\details{
It receives any number of parameters (`...`) that are automatically forwarded 
to the `startR::Start` function. See details in `?startR::Start`. The 
auxiliary functions used to define dimensions need to be called within the 
startR namespace (e.g. startR::indices(), startR::values(), startR::Sort(), 
startR::CircularSort(), startR::CDORemapper(), ...).
}
\examples{
\dontrun{
 sdates <- c('20101101', '20111101', '20121101')
 latmin <- 44
 latmax <- 47
 lonmin <- 6
 lonmax <- 9
 data <- CST_Start(dat = path,
                   var = 'prlr',
                   ensemble = indices(1:6),
                   sdate = sdates,
                   time = 121:151,
                   latitude = values(list(latmin, latmax)),
                   longitude = values(list(lonmin, lonmax)),
                   synonims = list(longitude = c('lon', 'longitude'),
                                   latitude = c('lat', 'latitude')),
                   return_vars = list(time = 'sdate',
                                      longitude = NULL, latitude = NULL),
                   retrieve = FALSE)
} 
\dontshow{
exp <- CSTools::lonlat_temp_st$exp
obs <- CSTools::lonlat_temp_st$obs
data <- CSTools::lonlat_prec
}
}
