% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.s2dv_cube.R
\name{print.s2dv_cube}
\alias{print.s2dv_cube}
\title{Print method for s2dv_cube objects}
\usage{
\method{print}{s2dv_cube}(x, ...)
}
\arguments{
\item{x}{An 's2dv_cube' object.}

\item{...}{Additional arguments of print function.}
}
\description{
This is an S3 method of the generic 'print' for the class 's2dv_cube'. When 
you will call 'print' on an 's2dv_cube' object, this method will display the 
content of the object in a clear and informative way.
}
\details{
The object will be displayed following 's2dv_cube' class conventions. The 
top-level elements are: 'Data', a multidimensional array containing the 
object's data; 'Dimensions', the dimensions of the array; 'Coordinates', the 
array coordinates that match its dimensions, explicit coordinates have an 
asterisk (*) at the beginning while index coordinates do not; and 
'Attributes', which contains all the metadata of the object. For more 
information about the 's2dv_cube', see \code{s2dv_cube()} and 
\code{as.s2dv_cube()} functions.
}
